move common tests into a separate package
