+++ wmsqlitemodernc/go.mod
@@ -2,7 +2,10 @@ module github.com/ThreeDotsLabs/watermill-sqlite/wmsqlitemodernc
 
 go 1.21
 
-require github.com/ThreeDotsLabs/watermill v1.4.6
+require (
+	github.com/ThreeDotsLabs/watermill v1.4.6
+	github.com/ThreeDotsLabs/watermill-sqlite/test v0.0.4
+)
 
 require (
 	github.com/cenkalti/backoff/v3 v3.2.2 // indirect
