+++ wmsqlitezombiezen/go.mod
@@ -4,7 +4,7 @@ go 1.21.0
 
 require (
 	github.com/ThreeDotsLabs/watermill v1.4.6
-	github.com/ThreeDotsLabs/watermill-sqlite/wmsqlitemodernc v0.0.4
+	github.com/ThreeDotsLabs/watermill-sqlite/test v0.0.4
 	github.com/google/uuid v1.6.0
 	zombiezen.com/go/sqlite v1.4.0
 )
